chore(deps): bump tower-http 0.7, jsonwebtoken 11, base64 0.23
Major-version upgrades of three deps; no change to rmcp-server-kit public API (these crate types are not re-exported).

tower-http 0.6->0.7: only call-site change is compression predicate SizeAbove::new, now u64 (was u16), wrapped with u64::from at the compression min-size site.

jsonwebtoken 10->11 (oauth feature): jwk::KeyAlgorithm is now non_exhaustive, already handled by the existing wildcard arm in jwk_algorithm; DEFAULT_PROVIDER, decode, and Validation unchanged.

base64 0.22->0.23: pinned default-features=false with the std feature so the new default-on simd-unsafe engine stays out of the tree (consistent with unsafe_code=forbid); Engine API unchanged.

cargo-vet exemptions recorded for the three new versions. base64 0.22 and tower-http 0.6 remain transitively via reqwest/rmcp, so cargo-deny emits non-failing duplicate-version warnings until upstreams update.

Verified green (--all-features): build, clippy -D warnings, test (543 passed), cargo deny check, cargo doc.

+++ src/transport.rs
@@ -1787,7 +1787,7 @@ where
     if config.compression_enabled {
         use tower_http::compression::Predicate as _;
         let predicate = tower_http::compression::DefaultPredicate::new().and(
-            tower_http::compression::predicate::SizeAbove::new(config.compression_min_size),
+            tower_http::compression::predicate::SizeAbove::new(u64::from(config.compression_min_size)),
         );
         router = router.layer(
             tower_http::compression::CompressionLayer::new()
